CFM-Rec/requirements.txt
tensorflow==2.15.0
pandas
numpy
pyyaml
tqdm
matplotlib